backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
elevenlabs
ffmpeg-python
python-dotenv
ffmpeg-python
SpeechRecognition pydub